Move pda/utils into utils directory
