+++ programs/perpetuals/tests/tests_suite/basic.rs
@@ -1,6 +1,3 @@
-use std::path::Path;
-
-use anchor_lang::prelude::{Rent, SolanaSysvar, UpgradeableLoaderState};
 use bonfida_test_utils::ProgramTestExt;
 use perpetuals::{
     instructions::{AddCustodyParams, InitParams, SetTestOraclePriceParams},
@@ -10,19 +7,14 @@ use perpetuals::{
         perpetuals::Permissions,
     },
 };
-use solana_program::{bpf_loader_upgradeable, stake_history::Epoch};
-use solana_program_test::{read_file, ProgramTest};
-use solana_sdk::{
-    account::Account,
-    signer::{keypair::Keypair, Signer},
-};
+use solana_program_test::ProgramTest;
+use solana_sdk::signer::{keypair::Keypair, Signer};
 
 use crate::{
     instructions::{
         test_add_custody, test_add_pool, test_init::test_init, test_set_test_oracle_price,
     },
-    pda::{self, get_test_oracle_account},
-    utils::{self, get_current_unix_timestamp},
+    utils::{self, add_perpetuals_program, get_current_unix_timestamp, get_test_oracle_account},
 };
 
 const ROOT_AUTHORITY: usize = 0;
@@ -68,50 +60,7 @@ pub async fn basic_test_suite() {
     };
 
     // Deploy the perpetuals program onchain as upgradeable program
-    {
-        // Deploy two accounts, one describing the program
-        // and a second one holding the program's binary bytes
-        let upgrade_authority = &keypairs[PERPETUALS_UPGRADE_AUTHORITY];
-
-        let mut program_bytes = read_file(
-            std::env::current_dir()
-                .unwrap()
-                .join(Path::new("../../target/deploy/perpetuals.so")),
-        );
-
-        let program_data_pda = pda::get_program_data_pda().0;
-
-        let program = UpgradeableLoaderState::Program {
-            programdata_address: program_data_pda,
-        };
-        let program_data = UpgradeableLoaderState::ProgramData {
-            slot: 1,
-            upgrade_authority_address: Some(upgrade_authority.pubkey()),
-        };
-
-        let serialized_program = bincode::serialize(&program).unwrap();
-
-        let mut serialzed_program_data = bincode::serialize(&program_data).unwrap();
-        serialzed_program_data.append(&mut program_bytes);
-
-        let program_account = Account {
-            lamports: Rent::default().minimum_balance(serialized_program.len()),
-            data: serialized_program,
-            owner: bpf_loader_upgradeable::ID,
-            executable: true,
-            rent_epoch: Epoch::default(),
-        };
-        let program_data_account = Account {
-            lamports: Rent::default().minimum_balance(serialzed_program_data.len()),
-            data: serialzed_program_data,
-            owner: bpf_loader_upgradeable::ID,
-            executable: false,
-            rent_epoch: Epoch::default(),
-        };
-
-        program_test.add_account(perpetuals::id(), program_account);
-        program_test.add_account(program_data_pda, program_data_account);
-    }
+    add_perpetuals_program(&mut program_test, &keypairs[PERPETUALS_UPGRADE_AUTHORITY]).await;
 
     let mut program_test_ctx = program_test.start_with_context().await;
