+++ programs/perpetuals/tests/utils/utils.rs
@@ -1,0 +1,102 @@
+use std::path::Path;
+
+use anchor_lang::prelude::*;
+use solana_program::{bpf_loader_upgradeable, stake_history::Epoch};
+use solana_program_test::{read_file, ProgramTest, ProgramTestContext};
+use solana_sdk::{account, signature::Keypair, signer::Signer};
+
+use super::get_program_data_pda;
+
+pub const ANCHOR_DISCRIMINATOR_SIZE: usize = 8;
+
+pub fn create_and_fund_account(address: &Pubkey, program_test: &mut ProgramTest) {
+    program_test.add_account(
+        *address,
+        account::Account {
+            lamports: 1_000_000_000,
+            ..account::Account::default()
+        },
+    );
+}
+
+pub fn find_associated_token_account(owner: &Pubkey, mint: &Pubkey) -> (Pubkey, u8) {
+    Pubkey::find_program_address(
+        &[
+            owner.as_ref(),
+            anchor_spl::associated_token::ID.as_ref(),
+            mint.as_ref(),
+        ],
+        &perpetuals::id(),
+    )
+}
+
+pub fn copy_keypair(keypair: &Keypair) -> Keypair {
+    Keypair::from_bytes(&keypair.to_bytes()).unwrap()
+}
+
+pub async fn get_account<T: anchor_lang::AccountDeserialize>(
+    program_test_context: &mut ProgramTestContext,
+    key: Pubkey,
+) -> T {
+    let account = program_test_context
+        .banks_client
+        .get_account(key)
+        .await
+        .unwrap()
+        .unwrap();
+
+    T::try_deserialize(&mut account.data.as_slice()).unwrap()
+}
+
+pub async fn get_current_unix_timestamp(program_test_context: &mut ProgramTestContext) -> i64 {
+    program_test_context
+        .banks_client
+        .get_sysvar::<solana_program::sysvar::clock::Clock>()
+        .await
+        .unwrap()
+        .unix_timestamp
+}
+
+// Deploy the perpetuals program onchain as upgradeable program
+pub async fn add_perpetuals_program(program_test: &mut ProgramTest, upgrade_authority: &Keypair) {
+    // Deploy two accounts, one describing the program
+    // and a second one holding the program's binary bytes
+    let mut program_bytes = read_file(
+        std::env::current_dir()
+            .unwrap()
+            .join(Path::new("../../target/deploy/perpetuals.so")),
+    );
+
+    let program_data_pda = get_program_data_pda().0;
+
+    let program = UpgradeableLoaderState::Program {
+        programdata_address: program_data_pda,
+    };
+    let program_data = UpgradeableLoaderState::ProgramData {
+        slot: 1,
+        upgrade_authority_address: Some(upgrade_authority.pubkey()),
+    };
+
+    let serialized_program = bincode::serialize(&program).unwrap();
+
+    let mut serialzed_program_data = bincode::serialize(&program_data).unwrap();
+    serialzed_program_data.append(&mut program_bytes);
+
+    let program_account = account::Account {
+        lamports: Rent::default().minimum_balance(serialized_program.len()),
+        data: serialized_program,
+        owner: bpf_loader_upgradeable::ID,
+        executable: true,
+        rent_epoch: Epoch::default(),
+    };
+    let program_data_account = account::Account {
+        lamports: Rent::default().minimum_balance(serialzed_program_data.len()),
+        data: serialzed_program_data,
+        owner: bpf_loader_upgradeable::ID,
+        executable: false,
+        rent_epoch: Epoch::default(),
+    };
+
+    program_test.add_account(perpetuals::id(), program_account);
+    program_test.add_account(program_data_pda, program_data_account);
+}
